+++ docs/history.rst
@@ -3,6 +3,11 @@
 Release History
 ===============
 
+`1.0.3`_ (13 Sep 2016)
+----------------------
+- Add a response ``content_length`` field, an ``accept`` tag (if set in request
+    headers), and a response ``content_type`` tag.
+
 `1.0.2`_ (13 Sep 2016)
 ----------------------
 - Don't use RequestHandler.reverse_url to get the endpoint pattern in the mixin
@@ -15,7 +20,8 @@ Release History
 ----------------------
 - Initial release
 
+.. _1.0.3: https://github.com/sprockets/sprockets-influxdb/compare/1.0.2...1.0.3
 .. _1.0.2: https://github.com/sprockets/sprockets-influxdb/compare/1.0.1...1.0.2
 .. _1.0.1: https://github.com/sprockets/sprockets-influxdb/compare/1.0.0...1.0.1
 .. _1.0.0: https://github.com/sprockets/sprockets-influxdb/compare/0.0.0...1.0.0
-.. _Next Release: https://github.com/sprockets/sprockets-influxdb/compare/1.0.2...master
+.. _Next Release: https://github.com/sprockets/sprockets-influxdb/compare/1.0.3...master